-fixed getmapping -getmapping has two parameters now -response is a nested JSON -changed getCost to be int instead of double -database file added

+++ src/main/java/com/HN/HiredKnight/EquipDecorator.java
@@ -18,7 +18,7 @@ public abstract class EquipDecorator implements Soldier {
     }
 
     @Override
-    public double getCost() {
+    public int getCost() {
         return tempSoldier.getCost();
     }
 }
