+++ src/main/java/com/HN/HiredKnight/Horse.java
@@ -14,7 +14,7 @@ public class Horse extends EquipDecorator {
     }
 
     @Override
-    public double getCost() {
+    public int getCost() {
         return tempSoldier.getCost() + 45;
     }
 }
